cleanup: moved to dokiHome component
